+++ mayan/apps/document_states/tests/test_workflow_transition_views.py
@@ -1,6 +1,8 @@
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
+from mayan.apps.events.tests.mixins import EventTestCaseMixin
 from mayan.apps.testing.tests.base import GenericViewTestCase
 
+from ..events import event_workflow_edited
 from ..literals import WIDGET_CLASS_TEXTAREA
 from ..models import WorkflowTransition
 from ..permissions import (
@@ -9,10 +11,7 @@ from ..permissions import (
 )
 
 from .literals import (
-    TEST_WORKFLOW_TRANSITION_FIELD_HELP_TEXT,
-    TEST_WORKFLOW_TRANSITION_FIELD_LABEL, TEST_WORKFLOW_TRANSITION_FIELD_NAME,
-    TEST_WORKFLOW_TRANSITION_FIELD_TYPE, TEST_WORKFLOW_TRANSITION_LABEL,
-    TEST_WORKFLOW_TRANSITION_LABEL_EDITED
+    TEST_WORKFLOW_TRANSITION_LABEL, TEST_WORKFLOW_TRANSITION_LABEL_EDITED
 )
 from .mixins import (
     WorkflowTestMixin, WorkflowTransitionEventViewTestMixin,
@@ -22,21 +21,27 @@ from .mixins import (
 
 
 class WorkflowTransitionViewTestCase(
-    WorkflowTestMixin, WorkflowViewTestMixin, WorkflowTransitionViewTestMixin,
-    GenericViewTestCase
+    EventTestCaseMixin, WorkflowTestMixin, WorkflowViewTestMixin,
+    WorkflowTransitionViewTestMixin, GenericViewTestCase
 ):
-    def test_create_test_workflow_transition_no_permission(self):
+    _test_event_object_name = 'test_workflow'
+
+    def test_workflow_transition_create_view_no_permission(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
+        self._clear_events()
 
         response = self._request_test_workflow_transition_create_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(WorkflowTransition.objects.count(), 0)
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
-    def test_create_test_workflow_transition_with_access(self):
+    def test_workflow_transition_create_view_with_access(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
+        self._clear_events()
 
         self.grant_access(
             obj=self.test_workflow, permission=permission_workflow_edit
@@ -58,11 +63,15 @@ class WorkflowTransitionViewTestCase(
             WorkflowTransition.objects.all()[0].destination_state,
             self.test_workflow_state_2
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event.verb, event_workflow_edited.id)
+        self.assertEqual(event.actor, self._test_case_user)
 
-    def test_delete_workflow_transition_no_permission(self):
+    def test_workflow_transition_delete_view_no_permission(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
+        self._clear_events()
 
         response = self._request_test_workflow_transition_delete_view()
         self.assertEqual(response.status_code, 404)
@@ -70,11 +79,14 @@ class WorkflowTransitionViewTestCase(
         self.assertTrue(
             self.test_workflow_transition in WorkflowTransition.objects.all()
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
-    def test_delete_workflow_transition_with_access(self):
+    def test_workflow_transition_delete_view_with_access(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
+        self._clear_events()
 
         self.grant_access(permission=permission_workflow_edit, obj=self.test_workflow)
 
@@ -84,11 +96,15 @@ class WorkflowTransitionViewTestCase(
         self.assertFalse(
             self.test_workflow_transition in WorkflowTransition.objects.all()
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event.verb, event_workflow_edited.id)
+        self.assertEqual(event.actor, self._test_case_user)
 
-    def test_edit_workflow_transition_no_permission(self):
+    def test_workflow_transition_edit_view_no_permission(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
+        self._clear_events()
 
         response = self._request_test_workflow_transition_edit_view()
         self.assertEqual(response.status_code, 404)
@@ -97,11 +113,14 @@ class WorkflowTransitionViewTestCase(
         self.assertEqual(
             self.test_workflow_transition.label, TEST_WORKFLOW_TRANSITION_LABEL
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
-    def test_edit_workflow_transition_with_access(self):
+    def test_workflow_transition_edit_view_with_access(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
+        self._clear_events()
 
         self.grant_access(
             obj=self.test_workflow, permission=permission_workflow_edit
@@ -115,22 +134,29 @@ class WorkflowTransitionViewTestCase(
             self.test_workflow_transition.label,
             TEST_WORKFLOW_TRANSITION_LABEL_EDITED
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event.verb, event_workflow_edited.id)
+        self.assertEqual(event.actor, self._test_case_user)
 
-    def test_workflow_transition_list_no_permission(self):
+    def test_workflow_transition_list_view_no_permission(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
+        self._clear_events()
 
         response = self._request_test_workflow_transition_list_view()
         self.assertNotContains(
             response=response, text=self.test_workflow_transition.label,
             status_code=404
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
-    def test_workflow_transition_list_with_access(self):
+    def test_workflow_transition_list_view_with_access(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
+        self._clear_events()
 
         self.grant_access(
             obj=self.test_workflow, permission=permission_workflow_view
@@ -141,6 +167,8 @@ class WorkflowTransitionViewTestCase(
             response=response, text=self.test_workflow_transition.label,
             status_code=200
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
 
 class WorkflowTransitionDocumentViewTestCase(
@@ -236,8 +264,7 @@ class WorkflowTransitionEventViewTestCase(
     WorkflowTestMixin, WorkflowTransitionEventViewTestMixin,
     GenericDocumentViewTestCase
 ):
-
-    def test_workflow_transition_event_list_no_permission(self):
+    def test_workflow_transition_event_list_view_no_permission(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
@@ -245,7 +272,7 @@ class WorkflowTransitionEventViewTestCase(
         response = self._request_test_workflow_transition_event_list_view()
         self.assertEqual(response.status_code, 404)
 
-    def test_workflow_transition_event_list_with_access(self):
+    def test_workflow_transition_event_list_view_with_access(self):
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
@@ -259,42 +286,21 @@ class WorkflowTransitionEventViewTestCase(
 
 
 class WorkflowTransitionFieldViewTestCase(
-    WorkflowTestMixin, WorkflowTransitionFieldTestMixin,
+    EventTestCaseMixin, WorkflowTestMixin, WorkflowTransitionFieldTestMixin,
     WorkflowTransitionFieldViewTestMixin, WorkflowTransitionViewTestMixin,
     GenericViewTestCase
 ):
+    _test_event_object_name = 'test_workflow'
+
     def setUp(self):
         super(WorkflowTransitionFieldViewTestCase, self).setUp()
         self._create_test_workflow()
         self._create_test_workflow_states()
         self._create_test_workflow_transition()
 
-    def test_workflow_transition_field_list_view_no_permission(self):
-        self._create_test_workflow_transition_field()
-
-        response = self._request_test_workflow_transition_field_list_view()
-        self.assertNotContains(
-            response=response,
-            text=self.test_workflow_transition_field.label,
-            status_code=404
-        )
-
-    def test_workflow_transition_field_list_view_with_access(self):
-        self._create_test_workflow_transition_field()
-
-        self.grant_access(
-            obj=self.test_workflow, permission=permission_workflow_edit
-        )
-
-        response = self._request_test_workflow_transition_field_list_view()
-        self.assertContains(
-            response=response,
-            text=self.test_workflow_transition_field.label,
-            status_code=200
-        )
-
     def test_workflow_transition_field_create_view_no_permission(self):
         workflow_transition_field_count = self.test_workflow_transition.fields.count()
+        self._clear_events()
 
         response = self._request_workflow_transition_field_create_view()
         self.assertEqual(response.status_code, 404)
@@ -303,13 +309,15 @@ class WorkflowTransitionFieldViewTestCase(
             self.test_workflow_transition.fields.count(),
             workflow_transition_field_count
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_workflow_transition_field_create_view_with_access(self):
         workflow_transition_field_count = self.test_workflow_transition.fields.count()
-
         self.grant_access(
             obj=self.test_workflow, permission=permission_workflow_edit
         )
+        self._clear_events()
 
         response = self._request_workflow_transition_field_create_view()
         self.assertEqual(response.status_code, 302)
@@ -318,11 +326,14 @@ class WorkflowTransitionFieldViewTestCase(
             self.test_workflow_transition.fields.count(),
             workflow_transition_field_count + 1
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event.verb, event_workflow_edited.id)
+        self.assertEqual(event.actor, self._test_case_user)
 
     def test_workflow_transition_field_delete_view_no_permission(self):
         self._create_test_workflow_transition_field()
-
         workflow_transition_field_count = self.test_workflow_transition.fields.count()
+        self._clear_events()
 
         response = self._request_workflow_transition_field_delete_view()
         self.assertEqual(response.status_code, 404)
@@ -331,11 +342,13 @@ class WorkflowTransitionFieldViewTestCase(
             self.test_workflow_transition.fields.count(),
             workflow_transition_field_count
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_workflow_transition_field_delete_view_with_access(self):
         self._create_test_workflow_transition_field()
-
         workflow_transition_field_count = self.test_workflow_transition.fields.count()
+        self._clear_events()
 
         self.grant_access(
             obj=self.test_workflow, permission=permission_workflow_edit
@@ -348,6 +361,75 @@ class WorkflowTransitionFieldViewTestCase(
             self.test_workflow_transition.fields.count(),
             workflow_transition_field_count - 1
         )
+        event = self._get_test_object_event()
+        self.assertEqual(event.verb, event_workflow_edited.id)
+        self.assertEqual(event.actor, self._test_case_user)
+
+    def test_workflow_transition_field_edit_view_no_permission(self):
+        self._create_test_workflow_transition_field()
+        workflow_transition_field_label = self.test_workflow_transition_field.label
+        self._clear_events()
+
+        response = self._request_workflow_transition_field_edit_view()
+        self.assertEqual(response.status_code, 404)
+
+        self.test_workflow_transition_field.refresh_from_db()
+        self.assertEqual(
+            workflow_transition_field_label,
+            self.test_workflow_transition_field.label
+        )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
+    def test_workflow_transition_field_edit_view_with_access(self):
+        self._create_test_workflow_transition_field()
+        workflow_transition_field_label = self.test_workflow_transition_field.label
+        self._clear_events()
+
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_edit
+        )
+
+        response = self._request_workflow_transition_field_edit_view()
+        self.assertEqual(response.status_code, 302)
+
+        self.test_workflow_transition_field.refresh_from_db()
+        self.assertNotEqual(
+            workflow_transition_field_label,
+            self.test_workflow_transition_field.label
+        )
+        event = self._get_test_object_event()
+        self.assertEqual(event.verb, event_workflow_edited.id)
+        self.assertEqual(event.actor, self._test_case_user)
+
+    def test_workflow_transition_field_list_view_no_permission(self):
+        self._create_test_workflow_transition_field()
+        self._clear_events()
+
+        response = self._request_test_workflow_transition_field_list_view()
+        self.assertNotContains(
+            response=response,
+            text=self.test_workflow_transition_field.label,
+            status_code=404
+        )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
+    def test_workflow_transition_field_list_view_with_access(self):
+        self._create_test_workflow_transition_field()
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_edit
+        )
+        self._clear_events()
+
+        response = self._request_test_workflow_transition_field_list_view()
+        self.assertContains(
+            response=response,
+            text=self.test_workflow_transition_field.label,
+            status_code=200
+        )
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
 
 class WorkflowInstanceTransitionFieldViewTestCase(